Refactor .conf SPT threshold switchover syntax
This commit replaces the switch_register_threshold and switch_data_threshold
with a unified syntax inspired by Cisco and Xorp PIM-SM implementations.

    spt_threshold [rate <KBPS> | packets <NUM> | infinity] [interval <SEC>]

The packets and infinity arguments are new, and the rate value is no
longer in bps, but in kbps.  Also, the default switchover is now based
on packets and triggers a switch to the SPT after receiving the first
encapsulated multicast frame on either the RP or the last-hop router.
The default interval has been increased from 20 to 100 sec, also like
Xorp does.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>